pyproject.toml
[project]
name = "project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "cloudinary>=1.44.1",
    "fastapi>=0.121.1",
    "imagekitio>=4.2.0",
    "psycopg2>=2.9.11",
    "pwdlib[argon2]>=0.3.0",
    "pydantic[email]>=2.12.4",
    "pyjwt>=2.10.1",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.20",
    "sqlalchemy>=2.0.44",
    "uvicorn>=0.38.0",
]
